--- Workouts/Workouts/View Controllers/CurrentWorkoutViewController.swift
@@ -1,19 +1,0 @@
-//
-//  CurrentWorkoutViewController.swift
-//  Workouts
-//
-//  Created by Дарья Воробей on 2/1/22.
-//
-
-import Foundation
-import UIKit
-
-class CurrentWorkoutViewController: UIViewController {
-    
-    @IBOutlet private var workoutCollectionView: UICollectionView!
-    @IBOutlet private var completeWorkoutButton: UIButton!
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
